feat: Implement exec abstraction layer over ECS and Kubernetes in Happy CLI (#623)
* feat: Implement exec abstraction layer over ECS and Kubernetes in Happy CLI

* Updated coverage, and moved executor to backend

* TTY shell

* Added better messaging

* Updated dependencies

* Update COVERAGE
