+++ cli/COVERAGE
@@ -1,1 +1,1 @@
-47.85
+45.51
