+++ cli/pkg/backend/aws/COVERAGE
@@ -1,1 +1,1 @@
-63.4
+54.9
