+++ cli/pkg/backend/aws/backend.go
@@ -76,6 +76,7 @@ type Backend struct {
 	// cached
 	username *string
 
+	executor       util.Executor
 	ComputeBackend interfaces.ComputeBackend
 }
 
@@ -88,6 +89,7 @@ func NewAWSBackend(
 	b := &Backend{
 		awsRegion:  aws.String("us-west-2"),
 		awsProfile: happyConfig.AwsProfile(),
+		executor:   util.NewDefaultExecutor(),
 	}
 
 	b.k8sClientCreator = func(config *rest.Config) (kubernetes.Interface, error) {
@@ -277,4 +279,8 @@ func (b *Backend) PrintLogs(ctx context.Context, stackName string, serviceName s
 
 func (b *Backend) RunTask(ctx context.Context, taskDefArn string, launchType config.LaunchType) error {
 	return b.ComputeBackend.RunTask(ctx, taskDefArn, launchType)
+}
+
+func (b *Backend) Shell(ctx context.Context, stackName string, service string) error {
+	return b.ComputeBackend.Shell(ctx, stackName, service)
 }
