+++ cli/pkg/backend/aws/ecs_test.go
@@ -32,7 +32,8 @@ func TestNetworkConfig(t *testing.T) {
 		SecurityGroups: sgs,
 		Services:       map[string]*config.RegistryConfig{},
 	}
-	networkConfig := backend.getNetworkConfig()
+	ecsBackend := ECSComputeBackend{Backend: &backend}
+	networkConfig := ecsBackend.getNetworkConfig()
 	r.NotNil(networkConfig)
 	r.Equal(len(subnets), len(networkConfig.AwsvpcConfiguration.Subnets))
 	r.Equal(len(sgs), len(networkConfig.AwsvpcConfiguration.SecurityGroups))
@@ -257,7 +258,8 @@ func TestEcsTasks(t *testing.T) {
 	r.NoError(err)
 	err = b.ComputeBackend.PrintLogs(ctx, "stack1", "frontend")
 	r.NoError(err)
-	taskId, err := b.getTaskID("arn:::::ecs/task/name/mytaskid")
+	ecsBackend := ECSComputeBackend{}
+	taskId, err := ecsBackend.getTaskID("arn:::::ecs/task/name/mytaskid")
 	r.NoError(err)
 	r.Equal("mytaskid", taskId)
 }
