+++ cli/pkg/backend/aws/interfaces/compute_backend.go
@@ -13,4 +13,5 @@ type ComputeBackend interface {
 	WriteParam(ctx context.Context, name string, val string) error
 	PrintLogs(ctx context.Context, stackName string, serviceName string, opts ...util.PrintOption) error
 	RunTask(ctx context.Context, taskDefArn string, launchType config.LaunchType) error
+	Shell(ctx context.Context, stackName string, serviceName string) error
 }
