+++ cli/pkg/orchestrator/orchestrator.go
@@ -3,8 +3,6 @@ package orchestrator
 import (
 	"context"
 	"fmt"
-	"os"
-	"os/exec"
 	"regexp"
 	"strings"
 	"time"
@@ -20,39 +18,16 @@ import (
 )
 
 type Orchestrator struct {
-	backend  *backend.Backend
-	executor util.Executor
-	dryRun   util.DryRunType
-}
-
-type container struct {
-	host          string
-	container     string
-	arn           string
-	taskID        string
-	launchType    string
-	containerName string
-}
-
-type TaskInfo struct {
-	TaskId     string `header:"Task ID"`
-	StartedAt  string `header:"Started"`
-	LastStatus string `header:"Status"`
+	backend *backend.Backend
+	dryRun  util.DryRunType
 }
 
 func NewOrchestrator() *Orchestrator {
-	return &Orchestrator{
-		executor: util.NewDefaultExecutor(),
-	}
+	return &Orchestrator{}
 }
 
 func (s *Orchestrator) WithBackend(backend *backend.Backend) *Orchestrator {
 	s.backend = backend
-	return s
-}
-
-func (s *Orchestrator) WithExecutor(executor util.Executor) *Orchestrator {
-	s.executor = executor
 	return s
 }
 
@@ -62,94 +37,7 @@ func (s *Orchestrator) WithDryRun(dryRun util.DryRunType) *Orchestrator {
 }
 
 func (s *Orchestrator) Shell(ctx context.Context, stackName string, service string) error {
-	clusterArn := s.backend.Conf().GetClusterArn()
-
-	serviceName := stackName + "-" + service
-	ecsClient := s.backend.GetECSClient()
-	//ec2Client := s.backend.GetEC2Client()
-
-	listTaskInput := &ecs.ListTasksInput{
-		Cluster:     aws.String(clusterArn),
-		ServiceName: aws.String(serviceName),
-	}
-
-	listTaskOutput, err := ecsClient.ListTasks(ctx, listTaskInput)
-	if err != nil {
-		return errors.Wrap(err, "error listing ecs tasks")
-	}
-
-	log.Println("Found tasks: ")
-	tablePrinter := util.NewTablePrinter()
-
-	describeTaskInput := &ecs.DescribeTasksInput{
-		Cluster: aws.String(clusterArn),
-		Tasks:   listTaskOutput.TaskArns,
-	}
-
-	describeTaskOutput, err := ecsClient.DescribeTasks(ctx, describeTaskInput)
-	if err != nil {
-		return errors.Wrap(err, "error describing ecs tasks")
-	}
-
-	containerMap := make(map[string]string)
-	var containers []container
-
-	for _, task := range describeTaskOutput.Tasks {
-		taskArnSlice := strings.Split(*task.TaskArn, "/")
-		taskID := taskArnSlice[len(taskArnSlice)-1]
-
-		startedAt := "-"
-
-		host := ""
-		if task.ContainerInstanceArn != nil {
-			host = *task.ContainerInstanceArn
-		}
-
-		if task.StartedAt != nil {
-			startedAt = task.StartedAt.Format(time.RFC3339)
-			containers = append(containers, container{
-				host:          host,
-				container:     *task.Containers[0].RuntimeId,
-				arn:           *task.TaskArn,
-				taskID:        taskID,
-				launchType:    string(task.LaunchType),
-				containerName: *task.Containers[0].Name,
-			})
-		}
-		containerMap[*task.TaskArn] = host
-		tablePrinter.AddRow(TaskInfo{TaskId: taskID, StartedAt: startedAt, LastStatus: *task.LastStatus})
-	}
-
-	tablePrinter.Flush()
-	// FIXME: we make the assumption of only one container in many places. need consistency
-	// TODO: only support ECS exec-command and NOT SSH
-	for _, container := range containers {
-		// This approach works for both Fargate and EC2 tasks
-		awsProfile := s.backend.Conf().AwsProfile()
-		log.Infof("Connecting to %s:%s\n", container.taskID, container.containerName)
-		// TODO: use the Go SDK and don't shell out
-		//       see https://github.com/tedsmitt/ecsgo/blob/c1509097047a2d037577b128dcda4a35e23462fd/internal/pkg/internal.go#L196
-		awsArgs := []string{"aws", "--profile", *awsProfile, "ecs", "execute-command", "--cluster", clusterArn, "--container", container.containerName, "--command", "/bin/bash", "--interactive", "--task", container.taskID}
-
-		awsCmd, err := s.executor.LookPath("aws")
-		if err != nil {
-			return errors.Wrap(err, "failed to locate the AWS cli")
-		}
-
-		cmd := &exec.Cmd{
-			Path:   awsCmd,
-			Args:   awsArgs,
-			Stdin:  os.Stdin,
-			Stderr: os.Stderr,
-			Stdout: os.Stdout,
-		}
-		log.Println(cmd)
-		if err := s.executor.Run(cmd); err != nil {
-			return errors.Wrap(err, "failed to execute")
-		}
-	}
-
-	return nil
+	return s.backend.Shell(ctx, stackName, service)
 }
 
 func (s *Orchestrator) TaskExists(ctx context.Context, taskType backend.TaskType) bool {
